refactor: use mixin instead
