+++ src/views/DocumentationTopic.vue
@@ -74,7 +74,6 @@
             :enableOnThisPageNav="enableOnThisPageNav"
           />
         </component>
-        <OnThisPageRegistrator :topicData="topicData" />
       </template>
     </CodeTheme>
   </div>
@@ -104,8 +103,7 @@ import StaticContentWidth from 'docc-render/components/DocumentationTopic/Static
 import { compareVersions, combineVersions } from 'docc-render/utils/schema-version-check';
 import { BreakpointName } from 'docc-render/utils/breakpoints';
 import { storage } from 'docc-render/utils/storage';
-import OnThisPageRegistrator
-  from 'docc-render/components/DocumentationTopic/OnThisPageRegistrator.vue';
+import OnThisPageRegistrator from '@/mixins/onThisPageRegistrator';
 import QuickNavigationStore from '../stores/QuickNavigationStore';
 
 const MIN_RENDER_JSON_VERSION_WITH_INDEX = '0.3.0';
@@ -115,7 +113,6 @@ export default {
   name: 'DocumentationTopicView',
   constants: { MIN_RENDER_JSON_VERSION_WITH_INDEX, NAVIGATOR_HIDDEN_ON_LARGE_KEY },
   components: {
-    OnThisPageRegistrator,
     Navigator,
     AdjustableSidebarWidth,
     StaticContentWidth,
@@ -125,7 +122,7 @@ export default {
     Nav: DocumentationNav,
     QuickNavigationModal,
   },
-  mixins: [performanceMetrics, onPageLoadScrollToFragment],
+  mixins: [performanceMetrics, onPageLoadScrollToFragment, OnThisPageRegistrator],
   data() {
     return {
       topicDataDefault: null,
